Changed thumbnail images, fixed errors.
Thumbnail pictures used to have aspect ratio of 16:10, making
the images disappear when the screen width was too wide.
Recaptured the images to be 16:9 (1280 x 720), fixing this issue.

Additionally, the site would fail to compile when running npm run build
due to faulty file paths for the background-vector.svg and unused import
statements in the about page.